Fix calc. of temperaturess from TFA wind sensor
The proposed change corrects the calculation of "temperature" and "chill" for TFA anemometers (wind sensors).  Previously, values < 0 were handled properly, but all values >0 were returned as 0 (due to the multiplication of the leading bit in the first line of each statement).

The proposed change duplicates the method used in the TempHum sensor further above in the code.

This change is operating correctly in my personal setup with HomeAssistant 0.42.3 (pyRFXtrx 0.17.0).

Please note that this has *not* been run through pylint, etc., and should be reviewed by someone more familiar with Python than me.